requirements.txt
mysql-connector-python>=8.2.0
python-dotenv>=1.0.0
openai>=1.12.0
python-telegram-bot>=20.7
typing-extensions>=4.8.0
